+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Watch-only wallets: support to delete credentials
 - Wallet onboard: automatic unique name creation support
 - Transaction summary dialog for hardware wallets
+- Jade firmware upgrade with OTA deltas
  
 ### Changed
 - Improve performance of login with hardware wallets